Rename header defining the interface for display functions.
